+++ InstaClone/src/main/java/com/skhkim/instaclone/chatting/service/ChatRoomService.java
@@ -5,5 +5,5 @@ import com.skhkim.instaclone.chatting.dto.ChatMessageDTO;
 
 public interface ChatRoomService {
     void updateLastCmId(ChatMessageDTO chatMessageDTO);
-    void updateUserNum(Long roomId, Integer addNum);
+    void updateUserNum(Long roomId, Long addNum);
 }
